+++ test/test.js
@@ -2069,6 +2069,32 @@ describe('i18next', function() {
         
       });
     
+      describe('extended - read key from inner content', function() {
+    
+        var resStore = {
+          dev: { translation: {  } },
+          en: { translation: {  } },            
+          'en-US': { translation: { 'simpleTest2': 'ok_from_en-US' } }
+        };
+        
+        beforeEach(function(done) {
+          setFixtures('<div id="container"><button id="testBtn" ' + opts.selectorAttr + '>simpleTest2</button></div>');
+    
+          i18n.init(i18n.functions.extend(opts, { 
+            resStore: resStore
+          }),
+            function(t) {
+              done(); 
+            });
+        });
+    
+        it('it should read key from inner content', function() {
+          $('#container').i18n(); // without option
+          expect($('#testBtn').text()).to.be('ok_from_en-US');
+        });
+        
+      });
+    
     });
   
   });